auth_service/requirements.txt
alembic==1.8.0
Flask-JWT-Extended==4.4.1
Flask-Migrate==3.1.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask==2.1.2
gevent==21.12.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
pydantic==1.9.0
python-dotenv==0.20.0
redis==4.3.3
SQLAlchemy==1.4.37
flask-restx==0.5.1
werkzeug==2.1.2
authlib==1.0.1
requests==2.28.1
flask-limiter==2.4.6
pytest==6.2.5
click~=8.1.3
authlib==1.0.1
opentelemetry-api==1.10.0
opentelemetry-sdk==1.10.0
opentelemetry-instrumentation-flask==0.29b1
opentelemetry-exporter-jaeger==1.10.0
jsonschema==4.5.1